pyproject.toml
[tool.poetry]
name = "yardgoat"
version = "0.1.0"
description = "Docker Execution Queue"
authors = ["Miller Wilt <miller@pyriteai.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.6"
boto3 = "^1.12.36"
docker = "^4.2.0"
toml = "^0.10.0"
attrs = "^19.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
mypy = "^0.770"
pylint = "^2.4.4"
pytest = "^5.4.1"
moto = "^1.3.14"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
